refactor: Extract getTracker to be able to override it
